Molecules cleanup moe (#67)
* formatting and fox new_molecule bug

* merge properties bug

* gui notebook and readme

* markdown test

* readme & gui demo

* gui readme update

* stash

* Support for display molecules from a dataframe

* stash

* prox server major upgrade and debug visuals

* prox server major upgrade and debug visuals

* prox server major upgrade and debug visuals

* stash

* Undid renaming gui-build to gui-build-proxy

* paths for routes

* acheived build

* merge conflicts mergibg with gui_api_moe #2

* Merge conflicts #3

* Merge wrapup

* Simplified/cleaned up JL_PROXY proxy code in gui_launcher

* Demo Notebook design

* patch launcher

* notebook updates

* stash

* gui-build & gui-build-proxy wip

* Updated gui build folders

* Updated gui build folders

* stash

* initial chemchat branch

* retrieving based on name before smils)

* turn spinner off models gen for api

* API IMPROVEMENTS

* API IMPROVEMENTS

* stash

* Update readme

* Update readme 1

* Update readme 2

* Update readme 3

* Update readme

* Update readme

* Update readme

* Update readme

* Update readme

* Update readme

* jupyter lab config sensing

* update for continer proxy base_url

* update for continer proxy base_url

* update for continer proxy base_url

* update for continer proxy base_url

* update for continer proxy base_url

* update for continer proxy base_url

* Update readme

* Update readme

* Update readme

* Update readme

* update for continer proxy base_url

* update for continer proxy base_url

* update for continer proxy base_url

* update for continer proxy base_url

* update for continer proxy base_url

* update for continer proxy base_url

* update for continer proxy base_url

* update for continer proxy base_url

* update for continer proxy base_url

* update for continer proxy base_url

* update for continer proxy base_url

* update for continer proxy base_url

* update for continer proxy base_url

* update for continer proxy base_url

* update for continer proxy base_url

* update for continer proxy base_url

* update for continer proxy base_url

* update for continer proxy base_url

* ollama fix

* ollama fix

* ollama fix

* update for continer proxy base_url

* demo update

* demo update

* debug set context

* debug set context

* debug set context

* debug set context

* debug set context

* debug set context

* debug set context

* debug set context

* debug set context

* debug set context

* debug set context

* debug set context

* debug set context

* debug set context

* debug set context

* Update readme

* Update readme

* Update readme

* Update readme

* Update readme

* Update readme

* Update readme

* Update readme

* Update readme

* Update readme

* Update readme

* Update readme

* Update readme

* Update readme

* Updated Notebook 'Common_Client_Intro.ipynb'

* Updated Notebook 'Common_Client_Intro.ipynb' #2

* Updated readme

* Updated readme

* Updated readme

* Updated readme

* Updated readme

* Updated readme

* Updated readme

* Updated readme

* add remove

* cross origin for flask apps

* cross origin for flask apps

* cross origin for flask apps

* cross origin for flask apps

* api decode info (#58) (#59)

* api decode info

* chore: lint


* reinstate commented out notebook line for importing molecules

* Fixed dataviewer submit with proxy url

* static data viewer

* Refactored doc generation + added auto-copy function to openad-docs repo

* Intro command will now also source the text from the docs folder, centralizing all documentation to a single source of truth

* Docs readme

* Docs readme

* Docs readme

* Docs readme

* Docs readme

* Docs readme

* Docs readme

* fix merge batch and doco

* update version

* update version

* formatting with lint

* Fixed display run header rendering issue

* Removed superfluous old gui-build-proxy files

* Fixed  command in Jupyter

* Removed banner.css which wasn't used

* Fixed notebook header layout issue ion wide screen, added remove run command

* Cleanup

* Removed trash

* Removed trash

* docs now also update main README.md file

* Updated documentation

* README.md now updated by documentation script + Analysis included in mol viewer

* Updated

* Removed gui-build and gui-build-proxy to overcome merge conflicts with main

* Restore gui-build folders from main to avoid conflicts

* Remove logger

* Macromnolecules base architecture (grammar, commands, functions, api)

* Established macromolecule file datastructure

* stash

* cif / pdb support

* Removed legacy files

* Added documentation

* Separated data formats in a separate file

* Cleanup

* cleanup

* Added search_fasta_sequence() to fetch protein by its FASTA string

* CIF meta data parsing + fetch mmol by identifier pipeline

* Moved massaging-for-human-readability of keys to frontend

* cleanup

* Removed biopython PDBParser, which was replaced with gemmi which reads pdb files as cif

* smol.json file support

* rename api routes mol to smol

* Renamed smols/mmols folders

* cleanup

* Fixed error with enrich api

* Ready for merge

* Removed mols2grid

* Removed mols2grid / molecule viewer

* Removed mols2grid / molecule viewer #2

* Removed debug loggers

* Poetry lock

* Updated GUI build

* Sorted MOL_PROPERTIES

* Mol functions refactor

* Mol functions refactor

* Cleanup of smol_functions.py

* CLI molecule printing improvements wip

* CLI molecule printing improvements wip

* CLI molecule printing improvements wip

* New gui build

* stash

* Improved display mol output

* Display mol output finetuning

* refactored @dopamine>>synonyms function and output, including changes to magic commands (to be tested)

* Fixed commands after refactor: add mol/remove mol/rename mol/export mol

* Fixed commands after refactor: export molecules/export molecule

* Fixed export mols/load mols commands and added support for sdf, smi and molset.json

* Fixed 'merge with pubchem' and created more robust merge molecule function

* Fixed language everywhere to refer to 'molecule working set' instead os molecule list/list of molecules/my mmolecules etc

* Fixed language everywhere to refer to 'molecule working set' instead os molecule list/list of molecules/my mmolecules etc

* Removed test code

* Consistent display of mol|molecule and molset|molecule-set in command docs

* Cleaned up molecule command documentation, halway done

* Cleaned up molecule grammar help, done

* Added user feedback for analysis results, added analysis to display mol output

* Updated help, fixed small issues after CLI testing

* Various cleanup

* Various small fixes

* various fixes

* openad_model_toolkit fix for canonical_smiles identifier

* Suppress RDKit warnings

* gui-build update

* Merge_mol improvement

* Fixed _enrich_with_pubchem_data (enrichg clause) for load mols from file

* Fix for merge mols data command

* Renamed verbose to uppercase

* fn_predict_reaction_batch_topn.py

* fix issue with basic molecule adding

* fix issue with basic molecule adding

* fix issue with basic molecule adding

* Fixed error with deprecated 'using' instead of 'from'

* Fix ignoring iupac_name as mol name when it's over 40 chars

* Move untracked test notebook to permanent location

* Fixed issue with viz data fetching

* Fixed minor issues with mere_mol function

* Added commands for viewing (GUI) and opening (OS) your workspace

* added a few new utility commands plus created GUI demo Notebook

* Updates to GUI demo Notebook

* Removed outdated Flask app demo

* Removed gui-build.zip that should not have been committed

* Fixed issue with pretty data, improved transposing of output table

* Removed static global print_width variable and replaced with heloper function so available print width is always calculated on the fly (so resizing terminal won't mess up output)

* Centralized all terminal width measurements so they call the same function with consistent result

* black

* version increment in metadata

---------

+++ openad/smols/smol_cache.py
@@ -8,6 +8,7 @@ import pickle
 from rdkit import Chem
 from openad.smols.smol_functions import canonicalize
 from openad.helpers.data_formats import ANALYSIS_RECORD
+from openad.helpers.output import output_success, output_error, output_warning
 
 
 CACHE_DIR = "/_result_cache/"
@@ -41,7 +42,7 @@ def save_result(result: dict, cmd_pointer) -> bool:
         inchi = Chem.rdinchi.MolToInchi(rdkit_mol)[0]
 
         inchikey = Chem.inchi.InchiToInchiKey(inchi)
-    except:
+    except Exception:  # pylint: disable=broad-except
         inchikey = result["smiles"]
         return False
 
@@ -51,10 +52,14 @@ def save_result(result: dict, cmd_pointer) -> bool:
 
 
 def _retrieve_results(smiles: str, cmd_pointer) -> list | bool:
-    """retrieves results from workspace cache"""
+    """
+    Retrieve results from workspace cache
+    """
+
     rdkit_mol = Chem.MolFromSmiles(smiles)  # pylint: disable=no-member
     inchi = Chem.rdinchi.MolToInchi(rdkit_mol)[0]
     inchikey = Chem.inchi.InchiToInchiKey(inchi)
+    print("inchikey", inchikey)
     results = []
     for i in glob.glob(
         _create_workspace_dir_if_nonexistent(cmd_pointer, CACHE_DIR) + inchikey + "-*.res",
@@ -66,14 +71,32 @@ def _retrieve_results(smiles: str, cmd_pointer) -> list | bool:
     return results.copy()
 
 
-def clear_results(cmd_pointer, inp) -> list | bool:
+def results_exist(cmd_pointer) -> bool:
+    """
+    Check if any results exist in the workspace cache.
+    """
+
+    if len(glob.glob(_create_workspace_dir_if_nonexistent(cmd_pointer, CACHE_DIR) + "*.res", recursive=True)) > 0:
+        return True
+    return False
+
+
+def clear_analysis(cmd_pointer, inp) -> list | bool:
     """clears results from workspace cache"""
+
+    # No alalysis results
+    if results_exist(cmd_pointer) is False:
+        return output_warning("Analysis results were already empty")
+
+    # Clear
     for i in glob.glob(
         _create_workspace_dir_if_nonexistent(cmd_pointer, CACHE_DIR) + "*.res",
         recursive=True,
     ):
         os.remove(i)
-    return True
+
+    # Success
+    return output_success("Analysis results have been cleared", return_val=True)
 
 
 def attach_results(smiles, cmd_pointer) -> bool:
@@ -84,31 +107,57 @@ def attach_results(smiles, cmd_pointer) -> bool:
     for mol in cmd_pointer.molecule_list:
         if "analysis" not in mol:
             mol["analysis"] = []
-        if mol["properties"]["canonical_smiles"] == smiles:
+        if mol["identifiers"]["canonical_smiles"] == smiles:
             for result in results:
-                if canonicalize(mol["properties"]["canonical_smiles"]) == canonicalize(result["smiles"].split("~")[0]):
+                if canonicalize(mol["identifiers"]["canonical_smiles"]) == canonicalize(result["smiles"].split("~")[0]):
                     if result not in mol["analysis"]:
                         mol["analysis"].append(result.copy())
     return True
 
 
-def attach_all_results(cmd_pointer, inp) -> bool:
-    """attaches a result to an existing molecule in the current working set, if it exists"""
-    i = 0
+def enrich_mws_with_analysis(cmd_pointer, inp) -> bool:
+    """
+    Attach the latest analysis results to a molecule in the current working set, if it present.
+    """
 
-    while i < len(cmd_pointer.molecule_list):
-        mol = cmd_pointer.molecule_list[i].copy()
-        if "analysis" not in mol:
-            mol["analysis"] = []
-        results = _retrieve_results(canonicalize(mol["properties"]["canonical_smiles"]), cmd_pointer)
+    mws = cmd_pointer.molecule_list
+    enriched_smols_count = 0
+
+    # No molecules
+    if len(mws) == 0:
+        return output_error("No molecules in the working set")
+
+    # No alalysis results
+    if results_exist(cmd_pointer) is False:
+        return output_error("No analysis results found in the workspace cache")
+
+    i = 0
+    while i < len(mws):
+        smol = mws[i].copy()
+        if "analysis" not in smol:
+            smol["analysis"] = []
+        results = _retrieve_results(canonicalize(smol["identifiers"]["canonical_smiles"]), cmd_pointer)
         for result in results:
-            if result not in mol["analysis"] and canonicalize(mol["properties"]["canonical_smiles"]) == canonicalize(
+            if result not in smol["analysis"] and canonicalize(smol["identifiers"]["canonical_smiles"]) == canonicalize(
                 result["smiles"].split("~")[0]
             ):
-                mol["analysis"].append(result)
-                cmd_pointer.molecule_list[i] = mol.copy()
+                smol["analysis"].append(result)
+                mws[i] = smol.copy()
+                enriched_smols_count = enriched_smols_count + 1
         i = i + 1
-    return True
+
+    # Success
+    if enriched_smols_count > 0:
+        return output_success(
+            [
+                f"{enriched_smols_count}/{len(mws)} molecules in your working set have been enriched with the latest analysis results",
+                "Run <cmd>show mols</cmd> to view the updated working set",
+            ]
+        )
+
+    # Error
+    else:
+        return output_error("No matching analysis results found for any of the molecules in your working set")
 
 
 def _write_analysis(result: dict, location):
